prepare for deployment

+++ Makefile
@@ -6,6 +6,6 @@ all:
 install:
 	./setup.py install
 	install -m 644 singsparrow-ii.service /etc/systemd/system/
-	install -m 644 singsparrow-ii.toml /etc/singsparrow-ii.toml
+	[ -f /etc/singsparrow-ii.toml ] || install -m 644 singsparrow-ii.toml /etc/singsparrow-ii.toml
 	systemctl daemon-reload || true
 	systemctl status singsparrow-ii.service --no-pager || true
